requirements.txt
dotenv
flask == 2.1.3
prometheus_client == 0.21.1
pyyaml == 5.4.1
waitress == 3.0.2
web3 == 7.6.1